ESCONF-13 use webpack 5
Upgrade to webpack v5. It would be so lovely if there were a way to
provide dev-peer-deps. Alas.

In fact, we already have a `eslint-config-stripe` `v5` branch for this
work, but the story is that we officially published `v5.4`, started work
on `v5.5` and published some intermediate changes to CI, then added some
breaking changes and published those to CI as `v6`. Thus, by inspecting
tags to figure out where we needed patch releases to CI of previous
releases, we found `5.4` but missed `5.5` because that latter was never
tagged and existed only in CI.

Refs ESCONF-13, STRWEB-4