ESCONF-18 bump stripes-webpack to v4 (#100)
Bump `@folio/stripes-webpack` to `v4` to stay in sync with the version
provided by `@folio/stripes-cli`. This makes me feel
`@folio/stripes-webpack` ought to be a peer, rather than a direct dep,
but I think that would only work in NPM (where peers are installed)
but not in yarn (where peers are ignored). Hmmmmmm.

Refs ESCONF-18, STCLI-203