ESCONF-21 update eslint-plugin versions (#103)
Update plugins to current versions to pickup features and bug-fixes,
e.g. recognizing state values in gDFSP as in-use.

The only major change here is to `no-only-tests` which has changed the
way block-scope matching works. We don't configure block-scope matching
so this is a non-issue for us.

Note that there are no changes to the config itself, either in `index.js` or to
the version of `eslint-config-airbnb` which could also introduce changes.
The idea is just to keep the parsers up to date. 

Refs ESCONF-21